kafka_manager/requirements.txt
confluent_kafka
kafka-python
psycopg2-binary
schedule